--- a/.gitignore
+++ b/.gitignore
@@ -3,5 +3,4 @@ _scratch*.jl
 venv
 build
 
-Manifest.toml
 copy_docs