--- a/src/carter-method/carter-boyer-lindquist.jl
+++ b/src/carter-method/carter-boyer-lindquist.jl
@@ -1,0 +1,11 @@
+@with_kw struct CarterBoyerLindquist{T}
+    @deftype T
+    "Photon Energy."
+    E = 1.0
+    "Black Hole Mass."
+    M = 1.0
+    "Black Hole Spin."
+    a = 0.0
+end
+
+export CarterBoyerLindquist