--- a/src/coordinates.jl
+++ b/src/coordinates.jl
@@ -79,7 +79,7 @@ Depending on whether the [`AbstractSpinDirection`](@ref) is `ContraRotating` or
 """
 @inline rms(M, a; ± = +) =
     M * (3 + Z₂(M, a) ± √((3 - Z₁(M, a)) * (3 + Z₁(M, a) + 2 * Z₂(M, a))))
-@inline function rms(s::BHSetup) 
+@inline function rms(s::BHSetup)
     metric = s.metric
     metric.a < 0.0 ? rms(metric.M, metric.a) : rms(metric.M, metric.a; ± = -)
 end