## New

- integrator methods now dispatch on (para)metric type of `BHSetup` for alternating between generic 2nd order, and Carter 1st order
- "old" method is now the `CarterBoyerLindquist` metric structure
- new integrator metrics defined in [ComputedGeodesicEquations.jl](https://github.com/astro-group-bristol/ComputedGeodesicEquations.jl), which defines, for `BoyerLindquist` and `EddingtonFinkelstein` the components of four-acceleration from the geodesic equation, as well as a constraint condition to ensure the velocity vectors are light-like

## Changes to existing
- the interface for creating `BHSetup` is now marginally different (code examples to follow)
- `GeodesicParams` is now used for 2nd order _only_, with 1st order method using `CarterGeodesicParams`

## Still needs
- documentation to be updated
- unit test need to be adjusted to instance different `AbstractGeodesicParams`
- somewhere to include SageMath recipes for generating Julia functions for different metrics (embedded?)

## Discussion

As anticipated, the 2nd order method solves the issues relating to dead pixels and sign changes, documented in the private repository.

There is numerous here to discuss, and new issues that need adjusting, which I will document as individual comments in this PR. The main overall advantage now is the ease with which different metrics and methods can be compared, and invites including even more integration methods for better comparison (e.g. that of Yang and Wang 2013).