@fjebaker 2021-11-13
## Benchmarks

The two methods (1st and 2nd order differential equations, hereafter "old" and "new") are benchmarked using
```julia
using GeodesicTracer
using BenchmarkTools

s = BHSetup(metric=CarterBoyerLindquist(M=1.0, a=0.0)) # method adjusted here

@benchmark calcgeodesic(rand(-4.0:4.0), rand(-4.0:4.0), s)
```
with integrator tolerances fixed to `1e-8`. The same was done with `a=1.0` and `a=-1.0` with negligible difference.

### Old method
```
BenchmarkTools.Trial: 852 samples with 1 evaluation.
 Range (min … max):  2.292 μs … 76.702 ms  ┊ GC (min … max): 0.00% … 89.02%
 Time  (median):     6.300 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   5.856 ms ±  5.415 ms  ┊ GC (mean ± σ):  5.42% ±  6.17%

  █                                     ▁▁                    
  █▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▃▂▂▂▂▃▇▇███▇▆▆▆▅▄▅▄▄▃▃▃▂▂▂▂▂▃ ▃
  2.29 μs        Histogram: frequency by time        9.32 ms <

 Memory estimate: 6.53 KiB, allocs estimate: 28.
```

### New method

```
BenchmarkTools.Trial: 849 samples with 1 evaluation.
 Range (min … max):  251.468 μs … 76.768 ms  ┊ GC (min … max): 0.00% … 89.49%
 Time  (median):       6.524 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):     5.883 ms ±  5.378 ms  ┊ GC (mean ± σ):  2.74% ±  4.37%

  █                                                             
  █▂▁▄▃▂▁▁▁▁▁▁▂▄▃▂▁▃▃▃▄▆▆▆▅▄▃▃▄▄▄▃▃▃▃▂▁▃▃▂▂▂▁▁▁▁▁▁▁▁▁▁▁▁▂▂▂▃▂▂ ▃
  251 μs          Histogram: frequency by time         17.5 ms <

 Memory estimate: 65.44 KiB, allocs estimate: 123.
 ```
 
 Note the number of allocations is approximates 6 times larger in the new method. This is to be expected, as the adaptive time stepping in Tsit5 is sensitive to stiffness. The 2nd order method is also calculating and storing velocity values, which have a steeper gradient close to to the event horizon, and so the time step is reduced and more steps are stored.
 
 ### Path comparison
 
Comparing the output of the two with `a=1.0`
 
 Old:
![old_test](https://user-images.githubusercontent.com/11492844/141647833-99e9b3cd-5522-4751-8602-b6151f7da256.png)
 
 New:
![new_test](https://user-images.githubusercontent.com/11492844/141647837-d7044885-874b-40e0-ab25-9c2c8ad1923f.png)

Note the curves are different, as the impact parameters are mapped a little different in both methods.

The two solutions are generally in good agreement! The 2nd order method handles the gradients close to the black hole considerably better than the 1st order method, but is slightly jumpier in interpolation in general: this can be adjusted with the tolerances, e.g. `1e-10`:

![new_test_e10](https://user-images.githubusercontent.com/11492844/141647929-9c8f1206-2d74-43cc-a44a-995b20df77f5.png)

Note, adjusting the tolerance, in this case by a factor of 100, is benchmarked 3 times slower with almost twice as many allocations.