@fjebaker 2021-11-13
## Shadow

Using the black hole shadow as another comparison, with integrator tolerance `1e-8`

```julia
using GeodesicTracer
using Plots

s = BHSetup(
    metric=CarterBoyerLindquist(M=1.0, a=0.0),  # method adjusted here
    fov_factor=9.0
)

const time = ValueFunction(
    (val, λ, u, p, d) -> λ
)

img = @time render(s, valuefunc=time)
heatmap(img, clims=(990.0, 1010.0))
```
Note that we are testing both the non-rotational case `a=0.0` and `a=1.0`: 

### Old method

Image sizes are relatively small here, so dead pixels are very improbable.

| Non-rotational: | Maximally-rotational: |
| - | - | 
| ![old_0 0_shadow](https://user-images.githubusercontent.com/11492844/141648186-6229f2ab-f327-4257-b2f7-43155cfb88a1.png) |  ![old_shadow](https://user-images.githubusercontent.com/11492844/141648196-3ed2c8bc-948d-474c-a4a8-f26e99b5a860.png) |

### New method

| Non-rotational: | Maximally-rotational: |
| - | - | 
| ![new_0 0_shadow](https://user-images.githubusercontent.com/11492844/141648218-4e11681f-20cd-4acf-9ec7-e8b9b64d86c7.png) | ![new_shadow](https://user-images.githubusercontent.com/11492844/141648238-90d829d2-996f-4a14-9777-6bdfe0767527.png) |

Obviously the rotational case is incorrect. There were a few factors which could explain this

- typo in geodesic equation generation
- velocity vectors are not normalised before being constrained
- impact parameters are being badly mapped when `|a| > 0.0`

It is also worth noting the difference in the methods as a test of the methods.

### Difference 

| Non-rotational: | Maximally-rotational: |
| - | - | 
| ![difference_0 0](https://user-images.githubusercontent.com/11492844/141648543-8af6b0f0-357c-4179-8d96-90e4eb200b84.png) | ![difference](https://user-images.githubusercontent.com/11492844/141648545-fe7feffc-1e6b-4b9e-a6c0-c4cc9d399c71.png) |

In the case of `a=0.0`, the difference is less than 1 part in 1000 at its worst, which is excellent. Surprisingly, the heuristic (discussed later) used for the parameter mapping seems to be very good for this case.

For `a=1.0`, we clearly see the geometric differences in the shadow calculated by the two methods.