@fjebaker 2021-11-13
## Geodesic paths visualised

To try and narrow down what the issue in the new method is, I plotted a subset of the geodesics as curves in 3d, hoping to discover if the impact parameters are badly mapped. If this were the case, it would be expected that the spacing between the light paths is irregular compared to the old method.

New method, for `a=0.0` for reference, with disk intersection shown:

<img width="640" alt="Screenshot 2021-11-13 at 01 41 34" src="https://user-images.githubusercontent.com/11492844/141648904-c01245d2-b6f3-47c5-af6a-7d1a734e2599.png">

Plotted are 3d geodesic paths for `a=1.0`, with no disk:

| Old | New |
| - | - |
| <img width="663" alt="Screenshot 2021-11-13 at 01 49 18" src="https://user-images.githubusercontent.com/11492844/141648823-8b0ba096-55f3-4a6b-92d7-2815ae3b31ca.png"> | <img width="664" alt="Screenshot 2021-11-13 at 01 45 11" src="https://user-images.githubusercontent.com/11492844/141648799-e1db651d-7a33-4702-ba6e-1feec7c9c04c.png"> |

From a slightly different angle:

| Old | New |
| - | - |
| <img width="653" alt="Screenshot 2021-11-13 at 01 56 49" src="https://user-images.githubusercontent.com/11492844/141648865-059a6bb1-7234-4441-b4ee-d25a033a099d.png"> | <img width="626" alt="Screenshot 2021-11-13 at 01 53 45" src="https://user-images.githubusercontent.com/11492844/141648884-b991a418-e57c-4cf6-9d01-e652224613d1.png"> |

It is a little tricky to capture in a screenshot, but in the new case, close to the black hole, there is force that pulls the geodesic paths towards the spin-axial poles.

This suggests some sort of calculation error, most likely a sign error, in the geodesic equation relating to the spin parameter `a`. However, if there is one, I have not yet been able to find it, so this is an on-going problem with the code.